Wait for rabbitmq before starting lma_collector
If we start lma_collector before the availability of rabbitmq cluster it
will fail to connect to the lma queues and then, it will fail to start.
It may take several long minutes before pacemaker starts the service.
So we need to be sure that rabbitmq cluster is up and running before
starting lma_collector.

Change-Id: Ia254b744f4173f64ee3ab8200b2896ecc412d06f