Replace per hypervisor metrics by aggregated hypervisor metrics
For now, we don't know how to display per hypervisor statistics (or
specially will be a mess with >100 compute nodes for example) and we
don't have the aggregated hypervisor metrics for the whole cloud.

In future, per hypervisor metrics could be re-enabled if relevant for
alerting purposes or whatever needs.

Change-Id: Iccbefc7c44df846a8351ff8d832e636474383c1b