Add parameter to give the host for the mod_status module
This change allows to pass a list of IPs that can be passed to the
mod_status apache module. On MOS 6.0 horizon is listening on the
management IP so the default parameter for mod_status (ie localhost) is
not working. With this patch we can pass an array of allow IPs.

Change-Id: I88644397c8c3074d5ab39f235406e5e137a6aff0