Lighten API checks
API checks are reduced to the bare minimum to avoid overloading the
endpoints.
We perform the most basic API test to determine API status, mainly by calling
the root path without authentication.
Those checks simply indicate the availability of the API but does not
provide health status (report OK even if the database is unreachable for
example).
Change-Id: I8caa31876f4534e1aace79a994176b281ec211b0