Add UI option to disable logs and notifications indexation
This change also enables the collection of logs and notifications even
if the events indexation is disabled. The reason is that useful metrics
are extracted from this data such as HTTP response times, resource
creation time and so on.
The term 'event' is used to reference both logs and OpenStack notifcations.

Change-Id: I28814129f745a8ef3ee44d3d3cf1cc267337b840