Fail early when the settings aren't valid
This change adds a post-deployment task that verifies that the plugin
configuration matches with the configuration of the environment. The
task checks that when using local Elasticsearch and InfluxDB nodes,
these nodes exist effectively in the deployed environment. The goal is
to stop the deployment as soon as possible if the configuration isn't
valid.

This couldn't be done in a pre-deployment task because Hiera isn't
available at this stage.

Change-Id: I318a214407a9e287f8376ec32907300bea554781