Set WriteQueueLimit parameters for collectd
This change sets the WriteQueueLimitHigh and WriteQueueLimitLow
for the collectd service. Without these settings, collectd might retain
all metrics in memory when the LMA collector is down or unreachable
which may lead to memory exhaustion eventually.

The same value is used for both parameters to avoid dropping random
metrics when the number of local metrics is between WriteQueueLimitLow
and WriteQueueLimitHigh.

We selected a default queue value of 10,000 items which should be enough
to avoid dropping metrics when collectd can reconnect to the LMA
collector after a reasonable amount of time.

Change-Id: I831b615169479bff4018aac6a398cf733dbafdb2