Tune the number of read threads for collectd
This change configures a sensible value for ReadThreads [1] depending on
the role of the node. A higher value is used for controller nodes
because their plugins perform more network I/O operations that may take
some time to complete.

We didn't change the WriteThreads setting because the default value
should be fine for all types of nodes. Anyhow the current collectd
Puppet module doesn't expose this parameter.

[1]
https://collectd.org/documentation/manpages/collectd.conf.5.shtml#readthreads_num

Change-Id: I95ef18b6ae6a317a16790d2a311d1540cd921eee