Increase the interval of service status computations
The interval is determined by the higher frequency used by collectd: 10s
Use a factor of 1.5 by default: hence statutes are computed every 15s.
By increasing the timer interval, all metrics have enough time to come
through the filter and this avoid to compute statutes with a subset of metric
and provide a partial status.

Change-Id: I53a93b5235953a80e0bcd4c54fdc0f2ae7b617b2