Add encoder + accumulator for InfluxDB v0.9
This change adds the initial implementation of the Heka output plugin
for InfluxDB v0.9.

First the InfluxDB accumulator catches all metric messages, buffers them
and injects a payload message containing the datapoints encoded using
the new InfluxDB line protocol. This message is injected either when the
number of datapoints reaches the configured limit (currently 100) or
after the configured time-out (currently 5 seconds). Eventually the
InfluxDB output plugin catches the message and sends the payload as-is
to the InfluxDB server thanks to the PayloadEncoder.

Change-Id: I5297c6e894f3f7b6d680a1fae2b73cbd81b60746
Implements: blueprint upgrade-influxdb-grafana