Add collectd notifications for Pacemaker resources
This change configures a collectd filter [1] that send notifications
reporting the states of the monitored Pacemaker resources. The
notification message is a JSON-encoded dictionary with 2 keys:
'resource' (the name of the Pacemaker resource) and 'value' (0 if the
resource isn't running locally, 1 otherwise).

The other collectd plugins may register their notification callback and
react on them as needed. But this will be implemented in another change.

[1] https://collectd.org/documentation/manpages/collectd.conf.5.shtml#filter_configuration

Change-Id: I40c34c903c147ce4770a57059d08fb5edc265100