Support for Ceph metrics with InfluxDB 0.9
This change updates the collectd plugins and decoder to support Ceph
metrics with InfluxDB 0.9.

It modifies the Base collectd plugin class used by the Ceph and RabbitMQ
plugins to leverage custom types instead of 'gauge' only. This allows to
have a cleaner definition for the collectd metrics and avoids encoding
too many information into the 'type_instance' field.

This patch also fixes a couple of bugs/inconsistencies in the Ceph
plugins inherited from the original plugins.

Change-Id: Ib070654ecb7128b647d6b9120d9d7dcc37a801bc
Implements: blueprint upgrade-influxdb-grafana