Collect Swift logs
This change configures the LMA collector to collect and parse the Swift
logs (except when the Ceph RADOS gateway is enabled). The parsing of
the logs is minimal for now because the Swift log format differs too
much from the format of the other OpenStack services.

Change-Id: I1bfafda1b85d9e40888b4992952b1150eb368007
Closes-Bug: #1480293
(cherry picked from commit c562fd22dbbc50340722d79c152c0eb76f5e6de0)