Avoid installing recommended packages by APT
We need to support the installation when using fuel-createmirror. It
means that we need to embed all packages that are required to install
collectd. As we don't want to embed recommended packages we need to
configure the packaging tool accordingly.

Closes-Bug: #1494186
Change-Id: I22ec782d747b3ccd0a0a4d4e39e7a94ba51a56e3
(cherry picked from commit ced36b62d11cdfdfba4d64c5e2a99ac52f179ee2)