Fix the firewall rule to comply with Zabbix
This changes the firewall rule securing the TCP input plugin of the LMA
collector. When Zabbix (or any other plugin configuring additional VIP
addresses) is deployed, the source address used by HAProxy may be any of
the management VIP addresses so the firewall rule should only check that
the source IP address is in the mangement subnet.

Change-Id: I84a27566d31098ce58121061aa713d26d4b51e58
Closes-Bug: #1506033