Catch buffer exceptions in Lua sandboxes
A Lua sandbox raises an exception when it tries to inject a message
larger than the configured output_limit value (default: 63KiB). The
same applies to the cjson library when trying to encode a Lua structure
resulting in a string larger than the same limit.

This change adds safe_* versions of the inject_message(),
inject_payload() and cjson.encode() functions. It also modifies the
existing Lua plugins to use the safe versions instead.

Change-Id: I7351783e51efa046d483921cb79e14279178a13a
Closes-Bug: #1504141