Refine disk alarms
This change modifies the disk alarms to be more fine-grained and provide
better insights. In particular, it creates a virtual cluster grouping
the MySQL nodes and configures alarms for the /var/lib/mysql partition.
When /var/lib/mysql gets full, only the status of the MySQL cluster will
be affected.

Change-Id: I0a893ffdf4e6cb8cd3847db72a7afa36995f1e6a