Enable buffering for HTTP output plugins
This change enables buffering for the Heka HTTP output plugin. We don't
enable buffering for Nagios outputs: it's ok to lose service checks
since Nagios is only interested by the current value.

It also adds validation for the buffering parameters because it
appeared during the tests that max_buffer_size needs to be greater than
max_file_size.

Change-Id: Ifbe98a2d649322f1213c4ecab71a95533f59510f