Add Puppet provider for managing Hiera sources
This change introduces a new Puppet provider that can add and remove
sources in the Hiera configuration. It replaces the previous mechanism
that used the file_line resource because it was too fragile and
couldn't cope with differences in YAML formatting.

Eventually this will be natively supported by the Fuel plugin framework
but it's not there yet.

Change-Id: Ie7563a9edf540bb81bdc30bc0200b7a25e25962f
Closes-Bug: #1527253
(cherry picked from commit ef2d5d8bb487bfe20997adf1dac44c5c43fa1800)