Remove lma_collector::collectd::service
This changes removes the lma_collector::collectd::service because it
is incompatible with lma_collector::collectd::python. The
incompatibility is related to both lma_collector::collectd::service
and lma_collector::collectd::python declaring a "collectd" service
resource. It is a mistake to have the lma_collector module declare
the "collectd" service, as the "collectd" is responsible for that.

Change-Id: I504753ebdf04fd77ab22be4d5caf94ea67db4397