Prevent MySQL conn error at collectd init
This change prevents a MySQL connection error visible in the collectd
logs at collectd start time.

The issue is fixed by making sure collectd uses TCP/IP rather the
Unix socket to connect to MySQL. This was the original intention,
but it's never worked as expected.

I took the opportunity to remove the "database" param to the
lma_collector::collectd::mysql class.  That param is not necessary as
any resource title can actually be used for
collectd::plugin::mysql::database.  Though I kept "nova" as the
resource title for compatibility reasons.

Change-Id: I20c9aab9b0c70a6de6c74ea9d2ea7dcc041fe5a5