Prevent MySQL conn error at collectd init
This change prevents a MySQL connection error visible in the collectd
logs at collectd start time.

The issue is fixed by making sure collectd uses TCP/IP rather the
Unix socket to connect to MySQL. This was the original intention,
but it's never worked as expected.

This change also removes the "database" param of the
lma_collector::collectd::mysql class because it's not necessary.

Change-Id: I20c9aab9b0c70a6de6c74ea9d2ea7dcc041fe5a5