Clean-up resources dealing with notifications
This change refactors the lma_collector Puppet module regarding the
processing of the OpenStack notification to get rid of the coupling with
Fuel. In particular, the configuration of the OpenStack services is
done in the external manifests since there was no point to have it in
lma_collector.

The change removes also workarounds that were necessary with older
versions of the plugin:

  - heat-engine is now managed as a regular service.
  - the can_exit flag is reverted back to false for the AMQP plugins

Change-Id: I39a2d25695449271b946ddcbca00cd8911dbdbb4