Change lma_collector::collectd::mysql host
This changes the host setting in lma_collector::collectd::mysql to
use the node name ($::hostname) rather than the management IP. Using
the management IP is the right way to connect to MySQL, but it is not
what we want for the Hostname tag in the metric values. Recent
versions of collectd include an Alias option that we could use here,
but the version of collectd we use does not have that option.

Closes-Bug: #1541818

Change-Id: Id19712a65db3327c18caafc779e524755458d6af