Make Elasticsearch health critical when an error occurs
Without this patch, if any error occurs when accessing the cluster
through the VIP the plugin doesn't emit any metric. This results to an
UNKNOWN status determined by the corresponding AFD.

With this patch, the cluster is explicitly declared 'CRITICAL' when there
has network error (eg the VIP is unavailable) or when the Elasticsearch
cluster responds with HTTP error.

Change-Id: Id69f466cd1c8872a051c130c40ea6a6ae315fd53