Update RabbitMQ log delimiter
This patch modifies the regex used by the rabbitmq splitter. It adds a
newline constant at the beginning of the regex to avoid to match the
first pattern of the log file. We observed that sometimes, the log
stream starts with a new line, so Heka finds a delimiter match on the
first two lines. But it does not have content for this log entry so it
may stop there.

Change-Id: Ib2208db842de9f30a95a514f0e51d00d693374e1
Closes-Bug: #1550222