Avoid forks when starting hekad
We must avoid to fork process when starting the Heka daemon because if
we kill the wrapper, the Heka daemon is detached and re-attach to the
init process. This can create several Heka daemon. To avoid this we need
to use exec and avoid the usage of sudo.

Fixes-bug: #1561109
(cherry picked from commit f0ba5e5077cdb63e7a4d2ccdd395465ee29fe34a)

Conflicts:
	deployment_scripts/puppet/modules/heka/spec/classes/heka_spec.rb

Change-Id: Idbfab2de92b993d1e5124de5bff44c4b09a88bb4