Update to task based deployment
MOS 9 comes with a new orchestration engine (so-called v2 or task-based
deployement) which reduces greatly the time necessary to deploy new
environments. This change updates the deployment tasks to be fully
compatible with the v2 engine since when a deployment mixes v1 and v2
tasks, the engine falls back to the v1 (slower) mode.

Note that we need to keep the backward compatibility with MOS 8 so the
deployment tasks still define the v1 attributes like 'requires' and
'required_for'.

Related-bug: #1558669

Change-Id: I92ede3808ad0b2bc9a39d4ec6560b1cc7ae256c3