Deserialize bulk metrics in the metric collector
This patch adds a custom decoder to be able to deserialize bulk_metrics.
The deserialization of bulk_metric message is configurable per Logger and
is disabled by default.
Currently, only the HTTP metrics are deserialized (Logger: aggregated_http_metrics_filter)
To deserialize several Loggers, a list of Loggers separated by space can
be configured for the sandbox.

blueprint: keystone-performance-monitoring

Change-Id: I08e8c816d889f4b0917d766b000fb25330eff174