Add roles with ['/.*/']
This patch adds a regex for roles because it has been modified into
fuel library and '*' is no longer supported by MOS 9. To keep the
compatibility with MOS 8 we still have the "role" definition. That means
that MOS 9 is using "roles" and ignores "role" whereas MOS 8 is using
"role" and ignores "roles".

Change-Id: Ib38c1b2bda634ed9e7fd21c88c535a125718f445