Fix tasks dependency when using remote backends
Previously the LMA collector had an explicit dependency on the last
Elasticsearch and InfluxDB tasks but this doesn't work when the
Elasticsearch/Kibana and InfluxDB-Grafana plugins aren't deployed in the same
environment as the collector plugin.

This change makes sure that the LMA collector plugin's tasks will only be
executed once the backends are fully functional by pinning all plugins to a
well-known task (update_hosts in that case).

Change-Id: I87a9067c5ace7acf3d676c7ad72d0b6632bf1869
Closes-Bug: #1573087
Depends-On: I830f024e0d1629ae10a8a504c4acbddf23d5c47e
Depends-On: I5bbc444eb51da4787b5be5cbe3aa0284ce6e1753