Increase the number of points per InfluxDB batch
This change improves the InfluxDB write performances by increasing to
500 the maximum number of points that are sent per InfluxDB request.
InfluxDB recommends to have a batch size of 5,000 but it cannot be the default
configuration value due to the fixed sized of Heka messages (256K currently)
which leads to silently discard metrics.
Note that the InfluxDB accumulator will flush the data either when
it holds 500 points or when it hasn't data for at least 5 seconds.

Co-Authored-By: Swann Croiset <scroiset@mirantis.com>

Change-Id: I7d238375dc0c231782983fc4901c9a32936fb08a
Partial-Bug: #1581369