Increase the number of points per InfluxDB batch
This change increases the maximum number of points to 500 that are sent in a
single request. InfluxDB recommends to have a batch size of 5,000
but it cannot be the default configuration value due to the limited
sized of heka messages (256K currently) which leads to silently discard
metrics.
Note that the InfluxDB accumulator will flush the data either when
it holds 500 points or when it hasn't data for at least 5 seconds.

This change reduces IOPS load on InfluxDB nodes.

Co-Authored-By: Swann Croiset <scroiset@mirantis.com>

Change-Id: I7d238375dc0c231782983fc4901c9a32936fb08a
Partial-Bug: #1581369