Increase the number of points per InfluxDB batch
This change increases the maximum number of points that are sent in a
single request. InfluxDB recommends to have a batch size of 5,000 so
this is now the default configuration value. Note that the InfluxDB
accumulator will flush the data either when it holds 5,000 points or
when it hasn't data for at least 5 seconds.

Change-Id: If07b7d285d216855997254952ca6d7511cff65ec
Partial-Bug: #1581369