Use Grafana credentials to monitor MySQL on dedicated environement
MySQL collectd plugins are configured to use Grafana credentials to
monitor the database when no controller is deployed within the
environement.
This assumes that MySQL is deployed with the detach database plugin.
If detach_database and influxdb_kibana roles are not deployed on same nodes,
collectd plugins, the MySQL monitoring is performed from the InfluxDB node.

Change-Id: Ida8d9f1995f87ca16b35ba59c3debd32f4410f97
Fixes-bug: #1578133