Update thresholds on free space for partitions
This patch sets the thresholds to 10% for critical alarms
on free space for partitions and 20% for warning alarms.
It doesn't change the threshold of the critical alarms for
elasticsearch fs because the modification for this one is
coming from a special commit that sets it to 15%. So we let
it as is.

Change-Id: I31eefb1dd5fcccbc7bc8db4cf21df8fbab12ea7b
Closes-Bug: #1587002