Add check metrics to collectd plugins
The check metrics are always emitted by Python plugins to express
the success or the failure of the metric collection.

Some plugins don't emit the check metric because it is redundant for them.
This concerns all the openstack_* plugins because the
check_openstack_api plugin has the same meaning than 'check' metrics.

The patch fixes also the error handling for the rabbitmq_info plugin.

Change-Id: Ic972e81ddd955e718bbc45b4b8502fa8ea3c2939