Add heartbeat metrics to Collectd plugins
The heartbeat metrics are always emitted by Python plugins to express
the success or the failure of the metric collection.
Some plugins skip the heartbeat because it is redondant for them.

The patch fixes also the error handling for the rabbitmq_info plugin.

All the openstack_* plugins are not impacted because the
check_openstack_api plugin is enough for the 'heartbeat' of OpenStack services.

Change-Id: Ic972e81ddd955e718bbc45b4b8502fa8ea3c2939