Add Pacemaker collectd plugin
This change adds a collectd plugin that gets metrics from the Pacemaker
cluster:

  - cluster's metrics
  - node's metrics
  - resource's metrics

Most of the metrics are only collected from the node that is the
designated controller except pacemaker_resource_local_active and
pacemaker_dc_local_active.

The plugin also removes the 'pacemaker_resource' plugin by providing the
exact same metrics and notifications for the other collectd plugins.

Finally the plugin is also installed on the standalone-rabbitmq and
standalone-database nodes if they are present.

Change-Id: I8b5b987704f69c6a60b13e8ea982f27924f488d1