Add monitoring of HDD errors
This change adds a filter plugin that monitors the kernel log messages
for hard drive errors and reports the number of errors per second
as 'hdd_errors_rate'. The filter is configured for all nodes,
irrespective of their roles. An alarm is also added that triggers
a WARNING alert when the metric value is greater than 0.

DocImpact

Change-Id: I485f5692a3e5facf0f7ea019ccdbd70683a7dd4e