Increase the Elasticsearch bulk size when required
The default bulk size for the Elasticsearch output plugin is too small
for 'slow' disks and the Elasticsearch cluster may always lag behind the
collectors.  This is particularly true for controllers that generate
OpenStack API logs in addition to Pacemaker logs or for all nodes when
the log level is set to debug.

In such cases, the flush_count parameter is increased to 100 (instead of
10 by default).

Change-Id: Ifdfbcb8ff0292f695dee4deab45560f126bde242
Closes-Bug: #1617211