Increase the Elasticsearch bulk size when required
The bulk size per default is too small for 'slow' disks and can lead to
never catch-up the amount of logs to index.
This is particularly true on clusters where are parsed logs of APIs and
Pacemaker.

The patch sets the flush_interval to 10s (vs 5s) and flush_count to 100
(vs 10).

Change-Id: Ifdfbcb8ff0292f695dee4deab45560f126bde242