Use netlink plugin instead of interface
This patch replaces the 'interface' collectd plugin by the 'netlink'
one. The 'netlink' plugin can get the same information than 'interface'
but it can also give other details like the number of dropped packets.
This new metrics will be used to generate alarms based on the number of
packets dropped by an interface.

Change-Id: I4e74e1e6c981a924946306192c296baef0404f95