Add 'no_data_policy' attribute to alarm definition
This allows to specify the behavior of the alarm when no datapoint have
been received in the defined window. The default value is to send an AFD
alert with 'UNKNOWN' to be backward-compatible with the existing alarms.

The value for the 'no_data_policy' attribute can either be a severity
level (okay, warning, critical, down, unknown) or 'skip' to avoid
sending the AFD metric.

This change is required to define relevant alarms against "sporadic"
metrics which aren't guaranteed to be received at periodic intervals
like for instance "resource creation time" metrics. It also needed to
support alarm definitions for cluster-wide metrics.

DocImpact

Change-Id: Iaa7fe63d321403ac2c5d4695a3346d912d3cc984