Add 'no_data_severity' attribute for AFD alarm definitions
This allows to specify the alarm's severity when no datapoint are
received. The default value is 'UNKNOWN' to be backward-compatible with
the existing alarms.

This change is required to define relevant alarms against "sporadic"
metrics, meaning metrics that aren't guaranteed to be received at
periodic intervals like for instance "resource creation time" metrics.

DocImpact

Change-Id: Iaa7fe63d321403ac2c5d4695a3346d912d3cc984