Fix issue when installing the OCF script
This patch moves the installation of the OCF script at the beginning of
the depoy_start to be sure that it is available when pacemaker starts
the collector resources. As it requires a configured hiera we also moved
the hiera task.

Change-Id: I90b4fa2a9038eaed0f1dcadb0f00713a1b2487b0
Closes-bug: #1575039
(cherry picked from commit 8bc835c74a51b85da03b4719ea35540908418277)