Decouple aggregator election from Pacemaker resource
The patch replaces the hard coded mechanism which is used to determine
which node is the current aggregator. This is currently used by GSE
filters and will be extended to some AFD filters.

The implementation uses the AFD framework to specify which particular
metric (and dimensions) is used for aggregator election.

Change-Id: I9ade5707b3fd7e9de0dfa48388cd2c5b5193988f